Implemented the SUS and temp driver method
Added methods to calculate the sus and it's score. Also added a driver method to test if the sus
was being calculated correct. By observation we have not seen in perform incorrectly. The driver
method is very temporary because it only will use the sus when moving and have no care for the lms
or the random scores.