Added methods to detect shortest with without knowledge of of current state.
If the agent has a perfect model of the environment but doesn't know what
state it's in, it can't follow an optimal path to the goal.  I've added a
method (which has two helper methods) that calculates (roughly) how many
steps it would take to reach the goal in this scenario.  Calculating the
exact minimum is NP-Hard and takes too long but this satisficing solution
is pretty close I think.  Some speedup could be had by adding memoization
but I think it's not worth the trouble at the moment.